+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add placeholder controllers for WebDAV HTTP methods.
 - Add placeholder controllers for file controllers.
 - Add PostElementController.
+- Refactor security tests of scenario 1.
 
 ## 0.0.5 - 2023-05-13
 ### Added